patch_ppns_in_databases.cc: Extended reporting for "report_only" flag

+++ cpp/pipelines/patch_ppns_in_databases.cc
@@ -94,55 +94,77 @@ void LoadMapping(MARC::Reader * const marc_reader,
 
 
 void PatchTable(DbConnection * const db_connection, const std::string &table, const std::string &column,
-                const std::vector<PPNsAndSigil> &old_ppns_sigils_and_new_ppns)
+                const std::vector<PPNsAndSigil> &old_ppns_sigils_and_new_ppns, const bool report_only)
 {
     const unsigned MAX_BATCH_SIZE(100);
 
-    db_connection->queryOrDie("BEGIN");
+    if (not report_only)
+        db_connection->queryOrDie("BEGIN");
 
     unsigned replacement_count(0), batch_size(0);
     for (const auto &old_ppn_sigil_and_new_ppn : old_ppns_sigils_and_new_ppns) {
         ++batch_size;
-        db_connection->queryOrDie("UPDATE IGNORE " + table + " SET " + column + "='" + old_ppn_sigil_and_new_ppn.new_ppn_
-                                  + "' WHERE " + column + "='" + old_ppn_sigil_and_new_ppn.old_ppn_ + "'");
-        replacement_count += db_connection->getNoOfAffectedRows();
-        if (batch_size >= MAX_BATCH_SIZE) {
-            db_connection->queryOrDie("COMMIT");
-            db_connection->queryOrDie("BEGIN");
+
+        const std::string where("WHERE " + column + "='" + old_ppn_sigil_and_new_ppn.old_ppn_ + "'");
+        if (report_only)
+            replacement_count += db_connection->countOrDie("SELECT count(*) AS replacement_count FROM " + table + " " + where, "replacement_count");
+        else {
+            db_connection->queryOrDie("UPDATE IGNORE " + table + " SET " + column + "='" + old_ppn_sigil_and_new_ppn.new_ppn_ + "' " + where);
+            replacement_count += db_connection->getNoOfAffectedRows();
+            if (batch_size >= MAX_BATCH_SIZE) {
+                db_connection->queryOrDie("COMMIT");
+                db_connection->queryOrDie("BEGIN");
+                batch_size = 0;
+            }
         }
     }
 
-    db_connection->queryOrDie("COMMIT");
+    if (report_only)
+        LOG_INFO("Would replace " + std::to_string(replacement_count) + " rows in " + table + ".");
+    else {
+        db_connection->queryOrDie("COMMIT");
+        LOG_INFO("Replaced " + std::to_string(replacement_count) + " rows in " + table + ".");
+    }
 
-    LOG_INFO("Replaced " + std::to_string(replacement_count) + " rows in " + table + ".");
 }
 
 
 void DeleteFromTable(DbConnection * const db_connection, const std::string &table, const std::string &column,
-                     const std::unordered_set<std::string> &deletion_ppns)
+                     const std::unordered_set<std::string> &deletion_ppns, const bool report_only)
 {
     const unsigned MAX_BATCH_SIZE(100);
 
-    db_connection->queryOrDie("BEGIN");
+    if (not report_only)
+        db_connection->queryOrDie("BEGIN");
 
     unsigned deletion_count(0), batch_size(0);
     for (const auto &deletion_ppn : deletion_ppns) {
         ++batch_size;
-        db_connection->queryOrDie("DELETE FROM '" + table + "' WHERE " + column + "='" + deletion_ppn + "'");
-        deletion_count += db_connection->getNoOfAffectedRows();
-        if (batch_size >= MAX_BATCH_SIZE) {
-            db_connection->queryOrDie("COMMIT");
-            db_connection->queryOrDie("BEGIN");
+
+        const std::string where("WHERE " + column + "='" + deletion_ppn + "'");
+        if (report_only)
+            deletion_count += db_connection->countOrDie("SELECT count(*) AS deletion_count FROM '" + table + "' " + where, "deletion_count");
+        else {
+            db_connection->queryOrDie("DELETE FROM '" + table + "' " + where);
+            deletion_count += db_connection->getNoOfAffectedRows();
+            if (batch_size >= MAX_BATCH_SIZE) {
+                db_connection->queryOrDie("COMMIT");
+                db_connection->queryOrDie("BEGIN");
+                batch_size = 0;
+            }
         }
     }
 
-    db_connection->queryOrDie("COMMIT");
-
-    LOG_INFO("Deleted " + std::to_string(deletion_count) + " rows from " + table + ".");
+    if (report_only)
+        LOG_INFO("Would delete " + std::to_string(deletion_count) + " rows from " + table + ".");
+    else {
+        db_connection->queryOrDie("COMMIT");
+        LOG_INFO("Deleted " + std::to_string(deletion_count) + " rows from " + table + ".");
+    }
 }
 
 
-void PatchNotifiedDB(const std::string &user_type, const std::vector<PPNsAndSigil> &old_ppns_sigils_and_new_ppns) {
+void PatchNotifiedDB(const std::string &user_type, const std::vector<PPNsAndSigil> &old_ppns_sigils_and_new_ppns, const bool report_only) {
     const std::string DB_FILENAME(UBTools::GetTuelibPath() + user_type + "_notified.db");
     if (not FileUtil::Exists(DB_FILENAME)) {
         LOG_INFO("\"" + DB_FILENAME + "\" not found!");
@@ -155,17 +177,23 @@ void PatchNotifiedDB(const std::string &user_type, const std::vector<PPNsAndSigi
     for (const auto &ppns_and_sigil : old_ppns_sigils_and_new_ppns) {
         if (db.keyIsPresent(ppns_and_sigil.old_ppn_)) {
             const std::string value(db.getValue(ppns_and_sigil.old_ppn_));
-            db.remove(ppns_and_sigil.old_ppn_);
-            db.addOrReplace(ppns_and_sigil.new_ppn_, value);
+
+            if (not report_only) {
+                db.remove(ppns_and_sigil.old_ppn_);
+                db.addOrReplace(ppns_and_sigil.new_ppn_, value);
+            }
             ++updated_count;
         }
     }
 
-    LOG_INFO("Updated " + std::to_string(updated_count) + " entries in \"" + DB_FILENAME + "\".");
+    if (report_only)
+        LOG_INFO("Would update " + std::to_string(updated_count) + " entries in \"" + DB_FILENAME + "\".");
+    else
+        LOG_INFO("Updated " + std::to_string(updated_count) + " entries in \"" + DB_FILENAME + "\".");
 }
 
 
-void DeleteFromNotifiedDB(const std::string &user_type, const std::unordered_set<std::string> &deletion_ppns) {
+void DeleteFromNotifiedDB(const std::string &user_type, const std::unordered_set<std::string> &deletion_ppns, const bool report_only) {
     const std::string DB_FILENAME(UBTools::GetTuelibPath() + user_type + "_notified.db");
     if (not FileUtil::Exists(DB_FILENAME)) {
         LOG_INFO("\"" + DB_FILENAME + "\" not found!");
@@ -177,12 +205,16 @@ void DeleteFromNotifiedDB(const std::string &user_type, const std::unordered_set
     unsigned deletion_count(0);
     for (const auto &deletion_ppn : deletion_ppns) {
         if (db.keyIsPresent(deletion_ppn)) {
-            db.remove(deletion_ppn);
+            if (not report_only)
+                db.remove(deletion_ppn);
             ++deletion_count;
         }
     }
 
-    LOG_INFO("Deleted " + std::to_string(deletion_count) + " entries from \"" + DB_FILENAME + "\".");
+    if (report_only)
+        LOG_INFO("Would delete " + std::to_string(deletion_count) + " entries from \"" + DB_FILENAME + "\".");
+    else
+        LOG_INFO("Deleted " + std::to_string(deletion_count) + " entries from \"" + DB_FILENAME + "\".");
 }
 
 
@@ -208,19 +240,19 @@ void AddPPNsAndSigilsToMultiMap(const std::vector<PPNsAndSigil> &old_ppns_sigils
 
 template<class SetOrMap, typename ProcessNotifieldDBFunc, typename ProcessTableFunc>
 void ProcessAllDatabases(DbConnection * const db_connection, const SetOrMap &set_or_map, const ProcessNotifieldDBFunc notified_db_func,
-                         const ProcessTableFunc table_func)
+                         const ProcessTableFunc table_func, const bool report_only)
 {
-    notified_db_func("ixtheo", set_or_map);
-    notified_db_func("relbib", set_or_map);
+    notified_db_func("ixtheo", set_or_map, report_only);
+    notified_db_func("relbib", set_or_map, report_only);
 
-    table_func(db_connection, "vufind.resource", "record_id", set_or_map);
-    table_func(db_connection, "vufind.record", "record_id", set_or_map);
-    table_func(db_connection, "vufind.change_tracker", "id", set_or_map);
+    table_func(db_connection, "vufind.resource", "record_id", set_or_map, report_only);
+    table_func(db_connection, "vufind.record", "record_id", set_or_map, report_only);
+    table_func(db_connection, "vufind.change_tracker", "id", set_or_map, report_only);
     if (VuFind::GetTueFindFlavour() == "ixtheo") {
-        table_func(db_connection, "ixtheo.keyword_translations", "ppn", set_or_map);
+        table_func(db_connection, "ixtheo.keyword_translations", "ppn", set_or_map, report_only);
         table_func(db_connection, "vufind.ixtheo_journal_subscriptions", "journal_control_number_or_bundle_name",
-                   set_or_map);
-        table_func(db_connection, "vufind.ixtheo_pda_subscriptions", "book_ppn", set_or_map);
+                   set_or_map, report_only);
+        table_func(db_connection, "vufind.ixtheo_pda_subscriptions", "book_ppn", set_or_map, report_only);
     }
 }
 
@@ -295,8 +327,6 @@ int Main(int argc, char **argv) {
             for (const auto &old_ppn_sigil_and_new_ppn : old_ppns_sigils_and_new_ppns)
                 LOG_INFO(old_ppn_sigil_and_new_ppn.old_ppn_ + " -> " + old_ppn_sigil_and_new_ppn.new_ppn_);
         }
-
-        return EXIT_SUCCESS;
     }
 
     if (old_ppns_sigils_and_new_ppns.empty())
@@ -310,12 +340,14 @@ int Main(int argc, char **argv) {
         return EXIT_SUCCESS;
     }
 
-    ProcessAllDatabases(&db_connection, old_ppns_sigils_and_new_ppns, PatchNotifiedDB, PatchTable);
+    ProcessAllDatabases(&db_connection, old_ppns_sigils_and_new_ppns, PatchNotifiedDB, PatchTable, report_only);
     AddPPNsAndSigilsToMultiMap(old_ppns_sigils_and_new_ppns, &already_processed_ppns_and_sigils);
-    MapUtil::SerialiseMap(ALREADY_SWAPPED_PPNS_MAP_FILE, already_processed_ppns_and_sigils);
+
+    if (not report_only)
+        MapUtil::SerialiseMap(ALREADY_SWAPPED_PPNS_MAP_FILE, already_processed_ppns_and_sigils);
 
 clean_up_deleted_ppns:
-    ProcessAllDatabases(&db_connection, deletion_ppns, DeleteFromNotifiedDB, DeleteFromTable);
+    ProcessAllDatabases(&db_connection, deletion_ppns, DeleteFromNotifiedDB, DeleteFromTable, report_only);
 
     return EXIT_SUCCESS;
 }
